Add FlexiViT models and weights, refactoring, push more weights * push all vision_transformer*.py weights to HF hub * finalize more pretrained tags for pushed weights * refactor pos_embed files and module locations, move some pos embed modules to layers * tweak hf hub helpers to aid bulk uploading and updating
