+++ timm/layers/helpers.py
@@ -10,7 +10,7 @@ import collections.abc
 def _ntuple(n):
     def parse(x):
         if isinstance(x, collections.abc.Iterable) and not isinstance(x, str):
-            return x
+            return tuple(x)
         return tuple(repeat(x, n))
     return parse
