+++ timm/models/_hub.py
@@ -111,14 +111,14 @@ def load_cfg_from_json(json_file: Union[str, os.PathLike]):
     return json.loads(text)
 
 
-def _download_from_hf(model_id: str, filename: str):
+def download_from_hf(model_id: str, filename: str):
     hf_model_id, hf_revision = hf_split(model_id)
     return hf_hub_download(hf_model_id, filename, revision=hf_revision)
 
 
 def load_model_config_from_hf(model_id: str):
     assert has_hf_hub(True)
-    cached_file = _download_from_hf(model_id, 'config.json')
+    cached_file = download_from_hf(model_id, 'config.json')
 
     hf_config = load_cfg_from_json(cached_file)
     if 'pretrained_cfg' not in hf_config:
@@ -145,21 +145,13 @@ def load_model_config_from_hf(model_id: str):
 
 def load_state_dict_from_hf(model_id: str, filename: str = 'pytorch_model.bin'):
     assert has_hf_hub(True)
-    cached_file = _download_from_hf(model_id, filename)
+    cached_file = download_from_hf(model_id, filename)
     state_dict = torch.load(cached_file, map_location='cpu')
     return state_dict
 
 
-def save_for_hf(model, save_directory, model_config=None):
-    assert has_hf_hub(True)
+def save_config_for_hf(model, config_path, model_config=None):
     model_config = model_config or {}
-    save_directory = Path(save_directory)
-    save_directory.mkdir(exist_ok=True, parents=True)
-
-    weights_path = save_directory / 'pytorch_model.bin'
-    torch.save(model.state_dict(), weights_path)
-
-    config_path = save_directory / 'config.json'
     hf_config = {}
     pretrained_cfg = filter_pretrained_cfg(model.pretrained_cfg, remove_source=True, remove_null=True)
     # set some values at root config level
@@ -170,11 +162,11 @@ def save_for_hf(model, save_directory, model_config=None):
     if isinstance(global_pool_type, str) and global_pool_type:
         hf_config['global_pool'] = global_pool_type
 
-    if 'label' in model_config:
+    if 'labels' in model_config:
         _logger.warning(
-            "'label' as a config field for timm models is deprecated. Please use 'label_name' and 'display_name'. "
+            "'labels' as a config field for timm models is deprecated. Please use 'label_name' and 'display_name'. "
             "Using provided 'label' field as 'label_name'.")
-        model_config['label_name'] = model_config.pop('label')
+        model_config['label_name'] = model_config.pop('labels')
 
     label_name = model_config.pop('label_name', None)
     if label_name:
@@ -194,6 +186,18 @@ def save_for_hf(model, save_directory, model_config=None):
 
     with config_path.open('w') as f:
         json.dump(hf_config, f, indent=2)
+
+
+def save_for_hf(model, save_directory, model_config=None):
+    assert has_hf_hub(True)
+    save_directory = Path(save_directory)
+    save_directory.mkdir(exist_ok=True, parents=True)
+
+    weights_path = save_directory / 'pytorch_model.bin'
+    torch.save(model.state_dict(), weights_path)
+
+    config_path = save_directory / 'config.json'
+    save_config_for_hf(model, config_path, model_config=model_config)
 
 
 def push_to_hf_hub(
